+++ src/jsonyx/test/test_syntax_error.py
@@ -108,11 +108,11 @@ def test_start_and_end_position(
         # No newline
         (9, "start-end", 0, 5, 1, "start-end", 6),
         #    ^^^^^                 ^^^^^
-        (8, "current\nnext", 0, 12, 1, "current", 8),
+        (8, "current\nnext", 0, 12, 1, "current", 5),
         #    ^^^^^^^^^^^^^              ^^^^^^^
-        (8, "current\rnext", 0, 12, 1, "current", 8),
+        (8, "current\rnext", 0, 12, 1, "current", 5),
         #    ^^^^^^^^^^^^^              ^^^^^^^
-        (8, "current\r\nnext", 0, 13, 1, "current", 8),
+        (8, "current\r\nnext", 0, 13, 1, "current", 5),
         #    ^^^^^^^^^^^^^^^              ^^^^^^^
 
         # Newline
@@ -120,11 +120,11 @@ def test_start_and_end_position(
         #           ^                   ^
         (8, "current", 7, 8, 8, "current", 9),
         #           ^                   ^
-        (8, "current\nnext", 7, 12, 8, "current", 9),
+        (8, "current\nnext", 7, 12, 8, "current", 5),
         #           ^^^^^^                     ^
-        (8, "current\rnext", 7, 12, 8, "current", 9),
+        (8, "current\rnext", 7, 12, 8, "current", 5),
         #           ^^^^^^                     ^
-        (8, "current\r\nnext", 7, 13, 8, "current", 9),
+        (8, "current\r\nnext", 7, 13, 8, "current", 5),
         #           ^^^^^^^^                     ^
 
         # At least one character