+++ docs/source/changelog.rst
@@ -25,7 +25,8 @@ jsonyx 2.0.0 (unreleased)
 - Added :func:`jsonyx.Manipulator`
 - Changed error for big integers to :exc:`jsonyx.JSONSyntaxError`
 - Fixed line comment detection
-- Fixed typo in error message
+- Fixed spelling of "commas" in error messages
+- Fixed ``end_offset`` of :exc:`jsonyx.JSONSyntaxError`
 - Improved documentation
 - Made :class:`tuple` JSON serializable
 - Merged ``item_separator`` and ``key_separator`` as ``separators`` for